Seperate Rank, Suit from Card class, modify CardDeck
Make Suit an enum type.

Create a Rank class to contain all the logic behind the rank of a card.

The Card class had private fields rank and suit as Strings, which were
prone to errors during value assignment. Change the Card class fields to
Suit enum and Rank class.

The CardDeck class wrapped a queue of Cards, which was not matching the
concept of a deck. Change the CardDeck to wrap a stack of cards and remove
unnecessary arrays of Ranks and Suits.