Add GameTable, modify Defense, Card, PlayerCircle
Defense class was huge and incomprehensible. Add a layer between
the Defense class and the cards on the table, by moving the
logic behind the validation of the cards to GameTable class.

Add GameTable class to contain the logic behind checking the
cards being put on the table.

The Card class returned a String in getRank() method. Change the
getRank() method to return a Rank object.

Add a next() method in PlayerCircle class.